pyproject.toml
[project]
name = "adk-python-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[eval]>=1.5.0",
    "langchain-community>=0.3.27",
    "langchain>=0.3.26",
    "toolbox-core>=0.3.0",
]

[tool.setuptools]
packages = ["tutorials"]

[dependency-groups]
dev = [
    "anyio>=4.9.0",
    "mypy>=1.16.1",
    "pytest-asyncio>=1.0.0",
    "pytest>=8.4.1",
    "ruff>=0.12.2",
    "pytest-tornasync>=0.6.0.post2",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.3",
    "twisted>=25.5.0",
]


[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
exclude = ["tutorials/wip"]

[tool.ruff]
target-version = "py311"
line-length = 127
indent-width = 4
exclude = [
    "*/__init__.py",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv/*",
    "local/*",
    "*/prompt.py",
    "*/wip",
]

[tool.ruff.lint]
ignore = ["B008"]
select = [
    "B",
    "E",
    "F",
    "I",
    "UP",
    "W",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 10

[tool.uv.workspace]
members = ["tutorials/tutorial_memory"]
